news_engine/requirements.txt
# GHOST News Engine Dependencies
# Основные зависимости для асинхронного сбора и анализа новостей

# HTTP клиент для асинхронных запросов
aiohttp==3.9.1
aiofiles==23.2.1

# Анализ настроения
nltk==3.8.1
textblob==0.17.1

# Работа с данными
pandas==2.1.4
numpy==1.24.3

# Конфигурация
pyyaml==6.0.1

# Логирование и мониторинг
structlog==23.2.0

# База данных
sqlite3  # Встроенный в Python
psycopg2-binary==2.9.9  # Для PostgreSQL (опционально)

# Обработка дат
python-dateutil==2.8.2

# Валидация данных
pydantic==2.5.2

# Кэширование
redis==5.0.1

# Уведомления
requests==2.31.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1

# Разработка
black==23.12.1
flake8==6.1.0
mypy==1.8.0

# Price Feed Engine Dependencies
# Дополнительные зависимости для сбора ценовых данных
websockets==12.0  # Для WebSocket соединений с биржами
ccxt==4.1.77  # Универсальная библиотека для работы с криптобиржами
